chore(import-request): address feedback
