--- src/routes/imports/find.ts
@@ -1,49 +1,0 @@
-import Joi from 'joi';
-import { ControllerGeneratorOptionsWithClient } from '../../comms/controller';
-import { TableQuery, EffectiveTableQuery, tableQuerySchemaGenerator } from '../../comms/table-controller';
-import { importSchema, Import } from '../../models/imports';
-
-type Query = TableQuery;
-
-type Request = {
-  query?: Query;
-} | undefined;
-
-type EffectiveQuery = EffectiveTableQuery;
-
-interface EffectiveRequest {
-  query: EffectiveQuery;
-}
-
-interface ResponseRow {
-  import: Import,
-}
-
-interface Response {
-  nextPageOffset: string | null;
-  rows: Array<ResponseRow>;
-}
-
-const controllerGeneratorOptions: ControllerGeneratorOptionsWithClient = {
-  method: 'get',
-  path: '/',
-  query: tableQuerySchemaGenerator(Joi.string().valid('hashId', 'name', 'state').default('name')),
-  right: { environment: 'USERS', supplier: 'ENVIRONMENT_ADMIN' },
-  response: Joi.object().keys({
-    nextPageOffset: Joi.string().allow(null).example(null).required()
-      .description('This is the last page if nextPageOffset is null'),
-    rows: Joi.array().items(Joi.object().keys({
-      import: importSchema.required(),
-    })).required(),
-  }),
-  description: 'Search through imports wihtin a monitoring environment',
-};
-
-export {
-  controllerGeneratorOptions,
-  Request,
-  EffectiveRequest,
-  Response,
-  Query,
-  ResponseRow,
-};
