--- src/routes/imports/routes.ts
@@ -1,13 +1,0 @@
-import * as generateTemplate from './generate-template';
-import * as add from './add';
-import * as get from './get';
-import * as deleteRoute from './delete';
-import * as find from './find';
-
-export {
-  generateTemplate,
-  add,
-  get,
-  deleteRoute as delete,
-  find,
-};
