+++ src/routes/index.ts
@@ -9,7 +9,7 @@ import CommandTypeRoute from './command-type';
 import DeviceRoute from './device';
 import DeviceTypeRoute from './device-type';
 import GraphRoute from './graph';
-import ImportsRoute from './imports';
+import ImportRequestRoute from './import-request';
 import IssueRoute from './issue';
 import IssueCommentRoute from './issue-comment';
 import ConnectivityThresholdRoute from './connectivity-threshold';
@@ -51,7 +51,7 @@ class Routes {
 
   graph: GraphRoute;
 
-  imports: ImportsRoute;
+  importRequest: ImportRequestRoute;
 
   issue: IssueRoute;
 
@@ -105,7 +105,7 @@ class Routes {
     this.deviceType = new DeviceTypeRoute(comms);
     this.environment = new EnvironmentRoute(comms);
     this.graph = new GraphRoute(comms);
-    this.imports = new ImportsRoute(comms);
+    this.importRequest = new ImportRequestRoute(comms);
     this.issue = new IssueRoute(comms);
     this.issueComment = new IssueCommentRoute(comms);
     this.connectivityThreshold = new ConnectivityThresholdRoute(comms);
@@ -141,7 +141,7 @@ export {
   DeviceTypeRoute,
   EnvironmentRoute,
   GraphRoute,
-  ImportsRoute,
+  ImportRequestRoute,
   IssueRoute,
   IssueCommentRoute,
   ConnectivityThresholdRoute,
